Complete Admin Layout (#5)
* Initial Layout & Sidebar (#1)

Created the initial layout and sidebar of the admin dashboard.

* Create Header and adjust layout (#2)

Created the header component and did some adjustments to the layout.
The adjustments consisted of spacing among the grid and columns and rows allocation

* Main content (#4)

* Initial Main Content Layout

Created the initial layout fo the main content area, to add more sections.

* Create Initial Main Content Area

Created the initial main content area
Made the initial layout for 2 sections:
- Posts
- Announcements and trending

* Fix Main Content Posts Layout

Fixed the layout of the main content's posts to be more responsive
The posts layout now changes according the the width of the screen

* Complete Initial Main Content Layout

Completed the initial main layout and styling

* Minor Styles Adjustments